- Updated README. - Added the code from original homework. - Put class NIM in the game_class file to clean up the main project file. Class Visual now creates a NIM object automatically.
